adding db access with django, deleting unused code, renaming microservice
