--- sentiment_tweet_job/.funcignore
@@ -1,16 +1,0 @@
-.git*
-.vscode
-local.settings.json
-.venv
-.DS_Store
-secrets*
-README.md
-__pycache__
-_twitter
-*.PublishSettings
-
-# Azurite artifacts
-__blobstorage__
-__queuestorage__
-__azurite_db*__.json
-.python_packages
