+++ sentiment_tweet_job/appsettings.Development.json
@@ -17,6 +17,15 @@
     "ManagedIdentity": {
       "Name": "twitter-bot-msi",
       "RestApiVersion": "2019-08-01"
+    },
+    "CognitiveServices": {
+      "AccountName": "twitter-bot-nlp"
+    },
+    "Databases": {
+      "CongressMember": {
+        "ServerName": "twitter-bot-sql-server",
+        "DatabaseName": "congress-member-db"
+      }
     }
   }
 }
