+++ sentiment_tweet_job/appsettings.Production.json
@@ -18,6 +18,15 @@
     "ManagedIdentity": {
       "Name": "twitter-bot-msi-prod",
       "RestApiVersion": "2019-08-01"
+    },
+    "CognitiveServices": {
+      "AccountName": "twitter-bot-nlp-prod"
+    },
+    "Databases": {
+      "CongressMember": {
+        "ServerName": "twitter-bot-sql-server-prod",
+        "DatabaseName": "congress-member-db-prod"
+      }
     }
   }
 }
