--- sentiment_tweet_job/config.azure.json
@@ -1,58 +1,0 @@
-{
-    "subscriptionId": "7cc90368-afd5-493e-9f7c-e5199bd142a5",
-    "resourceGroup": {
-        "name": "bot-rg",
-        "keyVault": {
-            "name": "twitter-bot-akv",
-            "secrets": [
-                "apiKey",
-                "apiKeySecret",
-                "bearerToken",
-                "clientId",
-                "clientSecret",
-                "accessToken",
-                "accessTokenSecret",
-                "twitterbotsaConnectionString"
-            ]
-        },
-        "functionApp": {
-            "name": "twitter-bot-func-app",
-            "functions": [
-                "get_users",
-                "get_sentiment",
-                "follow"
-            ]
-        },
-        "managedIdentity": {
-            "name": "twitter-bot-msi",
-            "clientId": "AppSettingMsiClientId",
-            "restApiVersion": "2019-08-01"
-        },
-        "storageAccount": {
-            "name": "twitterbotsa",
-            "containers": {
-                "follow": {
-                    "emptyFiles": []
-                },
-                "getsentiment": {
-                    "emptyFiles": [
-                        "current_index"
-                    ]
-                },
-                "getusers": {
-                    "emptyFiles": [
-                        "current.json",
-                        "missing.json",
-                        "found.json"
-                    ]
-                }
-            }
-        },
-        "cognitiveServices": {
-            "account": {
-                "name": "twitter-bot-nlp",
-                "endpoint": "https://twitter-bot-nlp.cognitiveservices.azure.com/"
-            }
-        }
-    }
-}
