--- sentiment_tweet_job/follow/function.json
@@ -1,14 +1,0 @@
-{
-    "disabled": false,
-    "scriptFile": "follow.py",
-    "bindings": [
-        {
-            "name": "inFound",
-            "type": "blobTrigger",
-            "dataType": "string",
-            "path": "getusers/found.json",
-            "connection": "AzureWebJobsStorage",
-            "direction": "in"
-        }
-    ]
-}
