--- sentiment_tweet_job/get_sentiment/function.json
@@ -1,36 +1,0 @@
-{
-    "disabled": false,
-    "scriptFile": "get_sentiment.py",
-    "bindings": [
-        {
-            "schedule": "0 0 19 * * *",
-            "name": "timer",
-            "type": "timerTrigger",
-            "direction": "in"
-        },
-        {
-            "name": "inFound",
-            "type": "blob",
-            "dataType": "string",
-            "path": "getusers/found.json",
-            "connection": "AzureWebJobsStorage",
-            "direction": "in"
-        },
-        {
-            "name": "inCurrentIndex",
-            "type": "blob",
-            "dataType": "string",
-            "path": "getsentiment/current_index",
-            "connection": "AzureWebJobsStorage",
-            "direction": "in"
-        },
-        {
-            "name": "outCurrentIndex",
-            "type": "blob",
-            "dataType": "string",
-            "path": "getsentiment/current_index",
-            "connection": "AzureWebJobsStorage",
-            "direction": "out"
-        }
-    ]
-}
