--- sentiment_tweet_job/get_users/function.json
@@ -1,60 +1,0 @@
-{
-    "disabled": false,
-    "scriptFile": "get_users.py",
-    "bindings": [
-        {
-            "schedule": "0 */10 * * * *",
-            "name": "timer",
-            "type": "timerTrigger",
-            "direction": "in"
-        },
-        {
-            "name": "inCurrent",
-            "type": "blob",
-            "dataType": "string",
-            "path": "getusers/current.json",
-            "connection": "AzureWebJobsStorage",
-            "direction": "in"
-        },
-        {
-            "name": "inMissing",
-            "type": "blob",
-            "dataType": "string",
-            "path": "getusers/missing.json",
-            "connection": "AzureWebJobsStorage",
-            "direction": "in"
-        },
-        {
-            "name": "inFound",
-            "type": "blob",
-            "dataType": "string",
-            "path": "getusers/found.json",
-            "connection": "AzureWebJobsStorage",
-            "direction": "in"
-        },
-        {
-            "name": "outCurrent",
-            "type": "blob",
-            "dataType": "string",
-            "path": "getusers/current.json",
-            "connection": "AzureWebJobsStorage",
-            "direction": "out"
-        },
-        {
-            "name": "outMissing",
-            "type": "blob",
-            "dataType": "string",
-            "path": "getusers/missing.json",
-            "connection": "AzureWebJobsStorage",
-            "direction": "out"
-        },
-        {
-            "name": "outFound",
-            "type": "blob",
-            "dataType": "string",
-            "path": "getusers/found.json",
-            "connection": "AzureWebJobsStorage",
-            "direction": "out"
-        }
-    ]
-}
