+++ sentiment_tweet_job/requirements.txt
@@ -1,11 +1,8 @@
-# DO NOT include azure-functions-worker in this file
-# The Python Worker is managed by Azure Functions platform
-# Manually managing azure-functions-worker may cause unexpected issues
-
+adal==1.2.7
+asgiref==3.6.0
 azure-ai-textanalytics==5.2.0b4
 azure-common==1.1.28
 azure-core==1.25.0
-azure-functions==1.11.2
 azure-identity==1.10.0
 azure-keyvault-secrets==4.4.0
 azure-mgmt-core==1.3.2
@@ -16,24 +13,44 @@ certifi==2022.12.7
 cffi==1.15.1
 charset-normalizer==2.1.0
 cryptography==39.0.1
+distlib==0.3.5
+Django==4.1.7
+docopt==0.6.2
+filelock==3.7.1
 idna==3.3
 isodate==0.6.1
 lxml==4.9.1
 msal==1.18.0
 msal-extensions==1.0.0
 msrest==0.7.1
+msrestazure==0.6.4
+mssql-django==1.2
 numpy==1.23.1
 oauthlib==3.2.2
 pandas==1.4.3
+pipreqs==0.4.11
+platformdirs==2.5.2
 portalocker==2.5.1
 pycparser==2.21
 PyJWT==2.4.0
+pyodbc==4.0.35
+PyQt3D==5.15.5
+PyQt5==5.15.7
+PyQt5-sip==12.11.0
+PyQtChart==5.15.6
+PyQtDataVisualization==5.15.5
+PyQtNetworkAuth==5.15.5
+PyQtPurchasing==5.15.5
+PyQtWebEngine==5.15.6
 python-dateutil==2.8.2
 pytz==2022.1
 requests==2.28.1
 requests-oauthlib==1.3.1
 six==1.16.0
+sqlparse==0.4.3
 StrEnum==0.4.8
 tweepy==4.7.0
 typing_extensions==4.3.0
 urllib3==1.26.11
+virtualenv==20.16.3
+yarg==0.1.9
