+++ sentiment_tweet_job/sentiment_tweet_job/enums/chamber.py
@@ -1,0 +1,5 @@
+from enum import Enum, auto
+
+class Chamber(Enum):
+    SENATE = auto()
+    HOUSE = auto()
