--- sentiment_tweet_job/twitter_bot_app/data/__init__.py
@@ -1,3 +1,0 @@
-from .model import Politician
-from .services import PoliticianDataService
-from .brokers import WikipediaDataBroker
