--- sentiment_tweet_job/twitter_bot_app/data/brokers/__init__.py
@@ -1,1 +1,0 @@
-from .wikipedia_data_broker import WikipediaDataBroker
