--- sentiment_tweet_job/twitter_bot_app/data/model/__init__.py
@@ -1,1 +1,0 @@
-from .politician import Politician
