--- sentiment_tweet_job/twitter_bot_app/data/services/__init__.py
@@ -1,1 +1,0 @@
-from .politician_data_service import PoliticianDataService
