--- sentiment_tweet_job/twitter_bot_app/enums/__init__.py
@@ -1,3 +1,0 @@
-from .politician_type import PoliticianType
-from .sentiment_analysis_category import SentimentScoreCategory
-from .sentiment import Sentiment
