--- sentiment_tweet_job/twitter_bot_app/enums/politician_type.py
@@ -1,11 +1,0 @@
-"""
-StrEnum defining allowed politician types
-"""
-from strenum import StrEnum
-
-class PoliticianType(StrEnum):
-    """
-    Defines enumeration for Representative, Senator politician types
-    """
-    REPRESENTATIVE = 'REPRESENTATIVE'
-    SENATOR = 'SENATOR'
