--- sentiment_tweet_job/twitter_bot_app/model/__init__.py
@@ -1,3 +1,0 @@
-from .sentiment_analysis import SentimentAnalysis
-from .twitter_account import TwitterAccount
-from .politician_twitter_account import PoliticianTwitterAccount
