--- sentiment_tweet_job/twitter_bot_app/model/politician_twitter_account.py
@@ -1,11 +1,0 @@
-from twitter_bot_app.data import Politician
-from .twitter_account import TwitterAccount
-
-class PoliticianTwitterAccount(TwitterAccount):
-    def __init__(
-        self,
-        politician: Politician,
-        **kwargs
-    ):
-        self.politician = politician
-        super().__init__(**kwargs)
