--- sentiment_tweet_job/twitter_bot_app/serialization/__init__.py
@@ -1,1 +1,0 @@
-from .encoder import Encoder
